Split Bible structures into modules
